// api/src/utils/respostas/codigos-resposta.d.ts
export declare const CODIGOS_ERRO: {
    ERRO_INTERNO: {
        codigo: number;
        mensagem: string;
    };
    DADOS_INCOMPLETOS: {
        codigo: number;
        mensagem: string;
    };
    DADO_TIPO_INVALIDO: {
        codigo: number;
        mensagem: string;
    };
    ID_INVALIDO: {
        codigo: number;
        mensagem: string;
    };
    CODIGO_INVALIDO: {
        codigo: number;
        mensagem: string;
    };
    MINUTOS_INVALIDOS: {
        codigo: number;
        mensagem: string;
    };
    HORAS_INVALIDAS: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_LISTAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_EXISTE_ERR: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_EXISTE_IG_ERR: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_EXISTE_TEL_ERR: {
        codigo: number;
        mensagem: string;
    };
    CLIENTE_N_EXISTE_ERR: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_LIST_ERR: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_CRIAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_EDITAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_EXCLUIR_ERR: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_N_EXISTE_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_LIST_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CRIAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_EDITAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_EXCLUIR_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_N_EXISTE_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_EXISTE_ENCOMENDA: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_LIST_ERR: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_CRIAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_EDITAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_EXCLUIR_ERR: {
        codigo: number;
        mensagem: string;
    };
    MATERIAL_N_EXISTE_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_LISTAR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_EXISTE_ERRO: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_N_EXISTE_ERRO: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_CRIAR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_EDITAR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_EXCLUIR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_MATERIAL_N_CORRESPONDE_ERRO: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_LISTAR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_EXISTE_ERRO: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_N_EXISTE_ERRO: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_CRIAR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_EDITAR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_EXCLUIR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    ENCOMENDA_MATERIAL_N_CORRESPONDE_ERRO: {
        codigo: number;
        mensagem: string;
    };
    RELATORIO_LISTAR_ERRO: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_CRIAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_EDITAR_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_EXCLUIR_ERR: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_EM_USO: {
        codigo: number;
        mensagem: string;
    };
    PRODUTO_CATEGORIA_N_EXISTE: {
        codigo: number;
        mensagem: string;
    };
};